app/detection/requirements.txt
numpy==1.21.6
onnxruntime==1.14.1
opencv-python-headless==4.9.0.80
Pillow # Use a version compatible with torchvision
torch==1.10.0
torchvision==0.11.1  # Adjust torchvision to match the compatible version with Pillow